requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
commonmark==0.9.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-admin-interface==0.19.1
django-colorfield==0.7.1
django-crispy-forms==1.14.0
django-filter==21.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
httpie==3.2.1
idna==3.3
importlib-metadata==4.11.4
Markdown==3.3.7
MarkupPy==1.14
multidict==6.0.2
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.1.1
Pygments==2.12.0
PySocks==1.7.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
typing-extensions==4.2.0
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
